+++ examples/data-sources/catalystcenter_wireless_rf_profile/data-source.tf
@@ -1,0 +1,3 @@
+data "catalystcenter_wireless_rf_profile" "example" {
+  id = "RF_Profile_1"
+}