+++ examples/resources/catalystcenter_wireless_rf_profile/import.sh
@@ -1,0 +1,1 @@
+terraform import catalystcenter_wireless_rf_profile.example "RF_Profile_1"