+++ docs/guides/changelog.md
@@ -18,6 +18,7 @@ description: |-
 - Add `catalystcenter_deploy_template` resource
 - Add `catalystcenter_lan_automation` resource and data source
 - Add `catalystcenter_wireless_enterprise_ssid` resource and data source
+- Add `catalystcenter_wireless_rf_profile` resource and data source
 
 ## 0.1.4